requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
aiomysql==0.2.0
